requirements.txt
requests
Django>=2.2
django-allauth>=0.39.1
psycopg2-binary
gunicorn
djangorestframework
Pillow
youtube_dl
sentry-sdk>=0.13.2
django-filter